Split ArtifactsCache out of artifacts.py
